Dialoguer moved
