+++ src/Services/FileManager.java
@@ -5,7 +5,6 @@
  */
 package Services;
 
-import View.Dialoguer;
 import java.io.File;
 
 /**
